+++ changelog.md
@@ -107,6 +107,7 @@
 * Fixed ore rocks not broken faster with pickaxes and added advancement
 * Fixed Echo Pillar advancement
 * Fixed sweet berries loot table
+* Fixed vanilla cutting the selected-slot highlight texture for ... reasons 
 
 ## 5.3.0.1-alpha
 * Removed Sweeping Edge enchantment